// src/swagger/model/roleForListViewModel.ts
/**
 * NSI.Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Permission } from './permission';


export interface RoleForListViewModel {
    roleInfoId?: number;
    name?: string;
    isActive?: boolean;
    manipulationLogId?: number;
    permissions?: Array<Permission>;
}